requirements.txt
apache-airflow[mysql]==2.6.0
pandas
mysql-connector-python
